#25 remove check of mrj.version for identifying Macintosh OS.
With Java 7 the system property mrj.version (which was used to identify
Mac OS in previous Java releases) doesn't return a non-null value on
Macs anymore.

Regarding to the note on
https://developer.apple.com/library/mac/technotes/tn2002/tn2110.html#MRJV
about the mrj.version, Macintoshs and Mac OS X are now identified by
checking if the system property os.name contains 'os x'.

This helps identifying Mac OS under Java 7 and following releases. But
it also removes support for identifying Mac OS 9. This shouldn't be
needed by anybody. But this commit should be changed if Mac OS 9 support
is needed!